Change regex for Imgur albums and galleries
Pictures from Imgur albums or galleries with query parameters and/or
fragments in the URL were not loading properly. This is because the
regex for both albums and galleries in the getImageTypeAndLink
function looked for an id that contained only letters and numbers
between the last "/" and the end of the URL string. Any URLs with
query parameters or fragments were then rejected because of the "?"
or "#" in the URL string.

The regex in the function was changed so that it looked for an id
containing only numbers and letters after the last "/", but does not
require it to be at the end of the string, so any query parameters or
fragments are ignored.